Generic update

+++ parser/types.go
@@ -2,13 +2,13 @@ package parser
 
 const (
 	//PTypeNone NA
-	PTypeNone int8 = -1
+	PTypeNone int = -1
 	//PTypeAddition Addition.
-	PTypeAddition int8 = 0
+	PTypeAddition int = 0
 	//PTypeMultiplication Multiplication.
-	PTypeMultiplication int8 = 1
+	PTypeMultiplication int = 1
 	//PTypeDivision Division.
-	PTypeDivision int8 = 2
+	PTypeDivision int = 2
 	//PTypeSubtraction Subtraction.
-	PTypeSubtraction int8 = 3
+	PTypeSubtraction int = 3
 )
